webrtc/RTCDtlsTransport-getRemoteCertificates.html

<!doctype html>
<meta charset="utf-8">
<title>RTCDtlsTransport.prototype.getRemoteCertificates</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="RTCPeerConnection-helper.js"></script>
<script src="RTCTransport-helper.js"></script>
<script>
  'use strict';

  // The following helper functions are called from RTCPeerConnection-helper.js:
  //  exchangeIceCandidates
  //  doSignalingHandshake

  // The following helper functions are called from RTCTransport-helper.js:
  //  waitConnectingPc
  //  createDtlsTransportsFromSctp
  //  createDtlsTransportsFromSenderReceiver

  function validateConnectedDtlsTransport(dtlsTransport) {
    const certs = dtlsTransport.getRemoteCertificates();

    assert_greater_than(certs.length, 0,
      'Expect DTLS transport to have at least one remote certificate when connected');

    for(const cert of certs) {
      assert_true(cert instanceof ArrayBuffer,
        'Expect certificate elements be instance of ArrayBuffer');
    }
  }

  function validateConnectingDtlsTransport(dtlsTransport) {
    if (dtlsTransport.state !== 'connected') {
      assert_array_equals(dtlsTransport.getCertificates(), [],
        'Expect DTLS certificates be initially empty until become connected');
    }
  }

  promise_test(t => {
    const pc1 = new RTCPeerConnection();
    const pc2 = new RTCPeerConnection();

    return createDtlsTransportsFromSenderReceiver(pc1, pc2)
    .then(([dtlsTransports1, dtlsTransports2]) => {
      const dtlsTransports = [...dtlsTransports1, ... dtlsTransports2];

      for (const dtlsTransport of dtlsTransports) {
        validateConnectingDtlsTransport(dtlsTransport);
      }

      return Promise.all([
        waitConnectingPc(pc1),
        waitConnectingPc(pc2)
      ])
      .then(() => {
        for (const dtlsTransport of dtlsTransports) {
          validateConnectedDtlsTransport(dtlsTransport);
        }
      });
    });
  }, 'RTP - connected DTLS transports should have valid remote certificates');

  promise_test(t => {
    const pc1 = new RTCPeerConnection();
    const pc2 = new RTCPeerConnection();

    return createDtlsTransportsFromSctp(pc1, pc2)
    .then(dtlsTransports => {
      for (const dtlsTransport of dtlsTransports) {
        validateConnectingDtlsTransport(dtlsTransport);
      }

      return Promise.all([
        waitConnectingPc(pc1),
        waitConnectingPc(pc2)
      ])
      .then(() => {
        for (const dtlsTransport of dtlsTransports) {
          validateConnectedDtlsTransport(dtlsTransport);
        }
      });
    });
  }, 'SCTP - connected DTLS transports should have valid remote certificates');

</script>
